orca-agent/requirements.txt
aiohttp==3.11.16
anthropic==0.49.0
colorama==0.4.6
fastapi==0.115.12
Flask==3.1.0
GitPython==3.1.44
gunicorn==23.0.0
kno-sdk==1.3.7
linguist==0.1.1
openai==1.68.2
prometheus-swarm==0.1.5
prometheus_client==0.21.1
PyGithub==2.6.1
python-dotenv==1.1.0
SQLAlchemy==2.0.39
sqlmodel==0.0.24
uvicorn==0.34.1
Werkzeug==3.1.3